build: simplify type definitions
