+++ static/scripts/types/audit.d.ts
@@ -61,3 +61,18 @@ export interface QuickImport {
   OWNER: string;
   REPO: string;
 }
+
+declare type TableIndexTypes = NumberConstructor | StringConstructor | BooleanConstructor | DateConstructor | ObjectConstructor | ArrayConstructor;
+interface TableIndex {
+  type: TableIndexTypes;
+  multiEntry?: boolean;
+  unique?: boolean;
+  default?: any;
+  ref?: string;
+}
+interface GoDBTableSchema {
+  [key: string]: TableIndex | TableIndexTypes;
+}
+export interface GoDBSchema {
+  [table: string]: GoDBTableSchema;
+}
