+++ static/scripts/types/index.ts
@@ -1,2 +1,1 @@
-export * from "./audit.types";
-export * from "./godb";
+export * from "./audit";
